chat/pyproject.toml
[project]
name = "chat"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "async-lru>=2.0.5",
    "beautifulsoup4>=4.13.4",
    "google-generativeai>=0.8.0",
    "langchain-google-genai>=2.1.0",
    "langchain>=0.3.27",
    "langgraph>=0.5.4",
    "markdownify>=1.1.0",
    "pillow>=11.2.1",
    "rich>=14.0.0",
]